LayoutTests/platform/chromium/fast/events/touch/gesture/pad-gesture-cancel-expected.txt

Tests basic use of GestureFlingCancel

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS actualWheelEventsOccurred is 0
PASS successfullyParsed is true

TEST COMPLETE
